+++ flynow-domain/src/main/java/com/flynow/domain/models/offer/OrganizerOfferDetails.java
@@ -19,7 +19,7 @@ public class OrganizerOfferDetails {
     private String imageFilename;
     private Integer eventOrganizerId;
     private String eventOrganizerName;
-    private Integer eventOrganizerRating;
+    private Float eventOrganizerRating;
     private String address;
     private String voivodeship;
     private LocalDateTime flightDate;
